+++ app/src/main/java/yokai/core/di/AppModule.kt
@@ -26,10 +26,6 @@ import eu.kanade.tachiyomi.source.SourceManager
 import eu.kanade.tachiyomi.util.chapter.ChapterFilter
 import eu.kanade.tachiyomi.util.manga.MangaShortcutManager
 import io.requery.android.database.sqlite.RequerySQLiteOpenHelperFactory
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.launch
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.protobuf.ProtoBuf
 import nl.adaptivity.xmlutil.XmlDeclMode
@@ -163,17 +159,4 @@ fun appModule(app: Application) = module {
 
     single { AndroidStorageFolderProvider(app) }
     single { StorageManager(app, get()) }
-}
-
-// REF: https://github.com/jobobby04/TachiyomiSY/blob/26cfb4811fef4059fb7e8e03361c141932fec6b5/app/src/main/java/eu/kanade/tachiyomi/di/AppModule.kt#L177C1-L192C2
-fun initExpensiveComponents(app: Application) {
-    // Asynchronously init expensive components on IO so cold start does not block the main thread.
-    CoroutineScope(SupervisorJob() + Dispatchers.IO).launch {
-        val koin = org.koin.core.context.GlobalContext.get()
-        koin.get<NetworkHelper>()
-        koin.get<SourceManager>()
-        koin.get<Database>()
-        koin.get<DownloadManager>()
-        koin.get<CustomMangaManager>()
-    }
 }